Fix panic during metrics manager startup
This fixes a regression introduced in #1876 where the driver would start
panicking on startup if `--http-endpoint` was specified. This was caused
by the metrics not being initialized anymore during startup. The
proposed fix involves using the `Reset` methods of the metrics object
instead of trying to redefine them each time they need to be reset.